backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.38.1
python-multipart==0.0.6
requests==2.31.0